feat(admin): implement CRUD with secure delete (CSRF + confirmation)

+++ app/src/Controller/AdminController.php
@@ -3,11 +3,9 @@
 namespace App\Controller;
 
 use App\Entity\Sentence;
-use App\Form\CommentType;
 use App\Form\SentenceType;
 use App\Repository\SentenceRepository;
 use Doctrine\ORM\EntityManagerInterface;
-use Dom\Comment;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
@@ -79,26 +77,20 @@ final class AdminController extends AbstractController
         ]);
     }
 
-
-    #[Route('/admin/{id}/delete',name: 'app_admin_delete')]
-    public function delete(Sentence $sentence,EntityManagerInterface $em): Response
-    {  
+    #[Route('/admin/{id}/delete', name: 'app_admin_delete')]
+    public function delete(Request $request, Sentence $sentence, EntityManagerInterface $em): Response
+    {
         $this->denyAccessUnlessGranted('ROLE_ADMIN');
-        $em->remove($sentence);
-        $em->flush();
-        $this-> addFlash('success', 'la phrase a bien été supprimée.');
+
+        $token = $request->request->get('_token');
+
+        if ($this->isCsrfTokenValid('delete' . $sentence->getId(), $token)) {
+            $em->remove($sentence);
+            $em->flush();
+
+            $this->addFlash('success', 'La phrase a bien été supprimée.');
+        }
 
         return $this->redirectToRoute('app_admin_index');
-
-
-
     }
-
-
-
-
-
-
-
-
 }
